Add script-auto-commit.sh and script-auto-commit.txt
This commit adds the script-auto-commit.sh and script-auto-commit.txt files. The script-auto-commit.sh file is a bash script that automates the process of committing and pushing changes to a Git repository. The script prompts the user to enter a directory, changes to that directory, executes a command, adds and pushes the changes to the repository, and then enters an infinite loop until the user presses the Esc key. The script-auto-commit.txt file is a sample text file used by the script.